requirements.txt
numpy==1.21.4
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.8.0
pip==22.0.4
